LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/referrer-policy-from-rules.https_4-4-expected.txt


FAIL unrecognized policies invalidate the rule assert_equals: must not be prefetched expected (undefined) undefined but got (string) "prefetch"
